+++ videos/instagram-reel/src/Root.tsx
@@ -4,6 +4,11 @@ import {
   CURRENT_STATE_FPS,
   ShellCurrentStateDemo,
 } from "./ShellCurrentStateDemo";
+import {
+  CURRENT_UI_LANDSCAPE_DURATION_FRAMES,
+  CURRENT_UI_LANDSCAPE_FPS,
+  ShellCurrentUiLandscape,
+} from "./ShellCurrentUiLandscape";
 import { ShellInstagramReel } from "./ShellInstagramReel";
 
 export const FPS = 30;
@@ -27,6 +32,14 @@ export const RemotionRoot = () => {
         fps={CURRENT_STATE_FPS}
         height={1280}
         width={720}
+      />
+      <Composition
+        id="ShellCurrentUiLandscape"
+        component={ShellCurrentUiLandscape}
+        durationInFrames={CURRENT_UI_LANDSCAPE_DURATION_FRAMES}
+        fps={CURRENT_UI_LANDSCAPE_FPS}
+        height={1080}
+        width={1920}
       />
     </>
   );